# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestion Medical",
        "description": "Documentation API pour la  gestion Medical",
        "contact": {
            "email": "diopawa425@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur API"
        }
    ],
    "paths": {
...
        "/api/personnels": {
            "get": {
                "tags": [
                    "personnels"
                ],
                "summary": "liste personnel",
                "description": "Liste de tous les personnels.",
                "operationId": "listPersonnel",
                "responses": {
                    "200": {
                        "description": "Données récupérées avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé, token invalide ou manquant."
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "personnels"
                ],
                "summary": "Créer un nouveau membre du personnel",
                "description": "Enregistre un nouveau membre du personnel dans la base de données.",
                "operationId": "createPersonnel",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Personnel"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Personnel"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/medecin-list": {
            "get": {
                "tags": [
                    "personnels"
                ],
                "summary": "liste des medecins",
                "description": "Liste de tous les medecins.",
                "operationId": "listMedecin",
                "responses": {
                    "200": {
                        "description": "Données récupérées avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé, token invalide ou manquant."
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
        "/api/medecins/service/{id}": {
            "get": {
                "tags": [
                    "personnels"
                ],
                "summary": "liste des medecins dans un service",
                "description": "Liste de tous les medecins dans un service.",
                "operationId": "listMedecinByService",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID service",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Données récupérées avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé, token invalide ou manquant."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Personnel": {
                "title": "Personnel",
                "description": "Schéma représentant un membre du personnel",
                "required": [
                    "name",
                    "prenom"
                ],
                "properties": {
                    "name": {
                        "description": "Nom du personnel",
                        "type": "string",
                        "example": "Diop"
                    },
                    "prenom": {
                        "description": "Prénom du personnel",
                        "type": "string",
                        "example": "Awa"
                    },
                    "telephone": {
                        "description": "Numéro de téléphone",
                        "type": "string",
                        "example": "+221772345678"
                    },
                    "adresse": {
                        "description": "Adresse",
                        "type": "string",
                        "example": "Dakar"
                    },
                    "datte_naissance": {
                        "description": "Date de naissance",
                        "type": "string",
                        "format": "date",
                        "example": "1985-10-15"
                    },
                    "lieu_naissance": {
                        "description": "Lieu de naissance",
                        "type": "string",
                        "format": "string",
                        "example": "1985-10-15"
                    },
                    "CNI": {
                        "description": "Numero carte d'identité",
                        "type": "string",
                        "format": "string",
                        "example": "154367548765"
                    },
                    "date_embauche": {
                        "description": "Mdate d'embauche",
                        "type": "string",
                        "format": "date",
                        "example": "1985-10-15"
                    },
                    "type_personnel_id": {
                        "description": "Type personnel associée",
                        "type": "integer",
                        "example": "1"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "service_id": {
                        "description": "Service affecté",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "title": "User",
                "description": "Schéma pour un utilisateur",
                "properties": {
                    "email": {
                        "description": "Email de l'utilisateur",
                        "type": "string",
                        "example": "user1@hopital.com"
                    },
                    "password": {
                        "description": "Mot de passe de l'utilisateur",
                        "type": "string",
                        "example": "password"
                    },
                    "roles": {
                        "description": "Role associée",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctumAuth": {
                "type": "http",
                "description": "Entrez le token Sanctum reçu après connexion.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}